changed userInfo type to NSDictionary*
git-svn-id: svn://orca.physics.unc.edu/Orca@11821 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Objects/Hardware/IPE/IPE SlowControl/ORIpeSlowControlModel.h
@@ -437,8 +437,8 @@
 - (void) setItemType:(int)aItemType;
 - (BOOL) viewItemName;
 - (void) setViewItemName:(BOOL)aViewItemName;
-- (unsigned) connectionHistoryCount;
-- (id) connectionHistoryItem:(unsigned)index;
+- (NSUInteger) connectionHistoryCount;
+- (id) connectionHistoryItem:(NSUInteger)index;
 - (NSString*) IPNumber;
 - (void) setIPNumber:(NSString*)aIPNumber;
 - (NSString*) ipNumberToURL;
@@ -451,7 +451,7 @@
 - (void) addItems:(NSArray*)aSensorPathArray;
 - (NSString*) itemKey:aUrl path:aPath;
 - (void) removeSet:(NSIndexSet*)aSetToRemove;
-- (unsigned) pendingRequestsCount;
+- (NSUInteger) pendingRequestsCount;
 - (id) pendingRequest:(id)aKey forIndex:(int)anIndex;
 
 - (void)      dumpSensorlist;
@@ -464,7 +464,7 @@
 
 
 #pragma mark ***Polling Cache
-- (unsigned)  pollingLookUpCount;
+- (NSUInteger)  pollingLookUpCount;
 - (NSString*) requestCacheItemKey:(int)anIndex;
 - (NSDictionary*) requestCacheItem:(int)anIndex;
